Fixes #54 Potential fix for Kotlin 1.8 JS issues
