+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,8 +1,8 @@
 object Versions {
     const val version = "0.12.0"
 
-    const val kotlin = "1.6.21"
-    const val serialization = "1.3.2"
+    const val kotlin = "1.8.0"
+    const val serialization = "1.5.0"
 
     const val mavenCentralPublish = "1.0.0-dev-3"
 }
